refactor: deduplicate the xcrun argument grammar into ProxyCore
ProxyServer.firstXcrunToolSelection and the private firstXcrunToolName
in RuntimeCoordinator+Readiness implemented the same flag-skipping
algorithm (with the same -sdk/--toolchain value-flag table) twice. The
grammar now lives once in ProxyCore.XcrunArguments, used by both the
readiness gate policy and the permission-dialog candidate derivation.
